backend/requirements.txt
# Python dependencies for Voice Cloning API

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.6

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0
requests>=2.31.0

# Background jobs
celery==5.4.0
redis==5.0.8

# Audio processing
pydub==0.25.1

# Environment variables
python-dotenv==1.0.0

# Authentication and database
fastapi-users[sqlalchemy]==13.0.0
sqlalchemy>=2.0.0
aiosqlite==0.19.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
email-validator==2.1.0
aiosmtplib==3.0.1

# Development dependencies (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
